siyologics-demo-platform/requirements.txt
Flask
Flask-Mail
mysql-connector-python
python-dotenv
razorpay
Werkzeug
pytz